Wrote the first function that can use a token to give you a tenant. (#1)
This is the beginnings of page token support and I now have to write some more code to make it
possible to use this code in a connect panel. But once that is possible I will have implimented the
required steps to have proper page token support in my Atlassian Connect applications.