Page token protected addition of pings. (#1)
Now youe can add pings that are protected by page tokens. Now the backend is in really good order
and we just need to make it so that the database gets populated with pings correctly. I am going to
start working on a really nice UI and UX for ping creation. That way we can feed it all through.