CHA-32: Adding the email into the hover message.
That way people can see the email adresses against the reminder and they can modify them to their
pleasure. That way you can always be sure which email address a reminder will be going to. It is a
little more dense but it is still the minimum information that you asked for.