CHA-31: Almost at the point where we can display all user reminders.
We want to make it such that all user reminders can be viewed in the one location. This should be
pretty easy to accomplish, and we should not have to worry about pagination at all since we will put
an upper bound on the number of reminders that a single user can create.