CHA-38: Working build and production images.
The good news is that I have created working build and production images with the production images
being only a fraction of the size of the build images. This is great as it will allow very
lightweight containers.

I want to see if I can get the dependencies down even more.