CHA-24: Started writing the purge handler.
This is the first step towards being able to purge customer data so that, when the customer stops
using our plugin (or JIRA) we stop sending them reminders. This will be really handy for our
customers but probably something that flies completely under the radar. Still needs to be done
however.