CHA-24: Purging customer data now partially works.
More work needs to be done in order to make the process complete. For example, it was a mistake to
make a new table called dormant tenant. I should just normalise that inside the tenant table with
null values. The purged_tenant table still makes a lot of sense and I will continue to use that for
this work.