CHA-38: Making it easier to build a production image.
It should be a one step repeatable process to build a production image from the source code and this
script should help us to accomplish that. Of course, thanks to cabal getting some freedom to choose
the dependencies it includes there is room for errors to sneak in between versions that are built at
different times.

There is not much that we can do about that and I think that there is not much that we should do
about that. This is why we should have verification integration tests.