MR-29: Cleaning up even more code.
Getting ready to do something with Routes.hs. That is the odd sheep out in the Atlassian Connect
code. I am beginning to think that the install handler will have to be implemented on a
per-application basis because the installed event could mean something different per application.
There is very little that we can do that will be common to every app.

The lifecycle response will be common to every application. So we can hold on to that.