MR-29: Moving non-connect code out into the application.
We have reached the point where most of the Atlassian connect code has been isolated. We have moved
as much of the Connect specific code as we can out and now we are at the point where we will want to
consider putting it out into libraries.

Current thoughts are:

 - atlassian-connect-descriptor
 - atlassian-connect-snaplet
 - atlassian-connect-postgresql

These three separate libraries are logical separation for our code.