MR-32: Caching static resources now works as expected.
We cache all static resources for a year and we provide redirects for all prior static resources
urls. Primarily so that we can always provide links to the latest version of that static resource.

The only thing now is that, every time we update the resources for My Reminders we need to bump the
version number of the addon as well. It is time to get into the habit of just bumping the version
number for every single release.